hacknova/hacknova_ps-main/backend/garak_scanner.py
import asyncio
import logging
import json
import uuid
import os
import subprocess
from typing import List, Dict, Any

logger = logging.getLogger(__name__)

def _load_available_probes() -> List[Dict[str, str]]:
    """Dynamically load probes from the garak library."""
    try:
        import garak._plugins
        cache = garak._plugins.PluginCache.instance()
        probes_cache = cache.get("probes", {})
        
        probes_dict = {}
        # Iterate over all probe classes in the cache
        for full_name, info in probes_cache.items():
            if not info.get("active", True):
                continue
            
            # Extract module name as the 'id' (e.g., 'dan' from 'probes.dan.Ablation_Dan_11_0')
            parts = full_name.split(".")
            if len(parts) < 2:
                continue
            
            probe_id = parts[1]
            if probe_id == "base":
                continue
                
            if probe_id not in probes_dict:
                probes_dict[probe_id] = {
                    "id": probe_id,
                    "name": probe_id.replace("_", " ").title(),
                    "desc": info.get("description", "").split("\n")[0] if info.get("description") else f"Garak {probe_id} probes"
                }
        
        return sorted(list(probes_dict.values()), key=lambda x: x["id"])
    except Exception as e:
        logger.error(f"Failed to dynamically load garak probes: {e}")
        return [
            {"id": "dan", "name": "Dan & Jailbreaks", "desc": "Various 'Do Anything Now' jailbreak techniques"},
            {"id": "promptinject", "name": "PromptInject", "desc": "Standardized prompt injection framework"},
            {"id": "web_injection", "name": "Web/XSS Injection", "desc": "Cross-site scripting and data exfiltration"}
        ]

# Dynamically populated list of Garak probes
AVAILABLE_PROBES = _load_available_probes()

class GarakScanner:
    def __init__(self, model_name: str, target_type: str = "huggingface.InferenceAPI", probes: List[str] = None, scan_id: str = None, log_callback=None):
        self.model_name = model_name
        self.target_type = target_type
        self.probes = probes or ["dan", "promptinject"]
        self.scan_id = scan_id or str(uuid.uuid4())
        self.output_file = f"/tmp/garak_report_{self.scan_id}"
        self.log_callback = log_callback
        
    async def execute_scan(self) -> List[Dict[str, Any]]:
        """
        Executes the NVIDIA garak scanner and parses the JSONL output.
        """
        logger.info(f"[garak_scanner] Starting garak scan for model {self.model_name} (Job: {self.scan_id}, Target: {self.target_type})")
        if self.log_callback:
            await self.log_callback(f"Starting garak scan for model {self.model_name} (Type: {self.target_type})...")

        # Prepare environment and command
        env = os.environ.copy()
        probe_args = ",".join(self.probes)
        
        # Ensure HF tokens are available for all generators (including secondary ones like atkgen)
        # Also set transformers/hub verbosity to avoid cluttering terminal logs
        hf_token = os.getenv("GARAK_HF_TOKEN")
        if hf_token:
            env["HF_TOKEN"] = hf_token
            env["HF_INFERENCE_TOKEN"] = hf_token
        
        env["TRANSFORMERS_VERBOSITY"] = "error"
        env["HUGGINGFACE_HUB_VERBOSITY"] = "error"

        # Determine actual generator to use. 
        # HuggingFace native InferenceAPI is currently broken with the new router (410/404 errors in Garak).
        # We use a workaround: use the 'openai' generator and redirect it to HF router.
        actual_target_type = self.target_type
        extra_args = []
        
        target_type_str = str(self.target_type).strip()
        target_type_lower = target_type_str.lower()
        
        logger.info(f"[garak_scanner] Target type processing: '{target_type_str}' -> lower: '{target_type_lower}'")

        # ONLY use the OpenAI-redirect workaround for InferenceAPI. 
        # If the user selects 'huggingface' (Local), we use Garak's native transformers pipeline.
        if "huggingface.inferenceapi" in target_type_lower:
            actual_target_type = "openai"
            # Use generator_options to set the HF router base URL (OpenAI compatible)
            # Correct base_url for HF Router is https://router.huggingface.co/v1
            extra_args = ["--generator_options", '{"base_url": "https://router.huggingface.co/v1"}']
            if hf_token:
                env["OPENAI_API_KEY"] = hf_token
            logger.info(f"[garak_scanner] Applied HF API workaround (actual_target_type: {actual_target_type})")
        elif "huggingface" == target_type_lower:
             # Native local transformers pipeline
             logger.info(f"[garak_scanner] Using native local huggingface generator (pipeline/transformers)")
        elif "openai" in target_type_lower:
            if os.getenv("GARAK_OPENAI_API_KEY"):
                env["OPENAI_API_KEY"] = os.getenv("GARAK_OPENAI_API_KEY")
        elif "nim" in target_type_lower:
            if os.getenv("GARAK_NIM_API_KEY"):
                env["NIM_API_KEY"] = os.getenv("GARAK_NIM_API_KEY")
                env["OPENAI_API_KEY"] = os.getenv("GARAK_NIM_API_KEY") # NIM often uses OpenAI SDK
        elif "replicate" in target_type_lower:
            if os.getenv("GARAK_REPLICATE_API_TOKEN"):
                env["REPLICATE_API_TOKEN"] = os.getenv("GARAK_REPLICATE_API_TOKEN")

        cmd = [
            "python3", "-m", "garak",
            "--model_type", actual_target_type,
            "--model_name", self.model_name,
            "--probes", probe_args,
            "--report_prefix", self.output_file
        ] + extra_args
        
        logger.info(f"[garak_scanner] Executing: {' '.join(cmd)}")
        if self.log_callback:
            await self.log_callback(f"Executing CLI: {' '.join(cmd)}")
        
        # Run subprocess asynchronously and capture output
        process = await asyncio.create_subprocess_exec(
            *cmd,
            env=env,
            stdout=asyncio.subprocess.PIPE,
            stderr=asyncio.subprocess.STDOUT
        )
        
        # Read logs line by line and stream them
        while True:
            line = await process.stdout.readline()
            if not line:
                break
            log_line = line.decode().strip()
            if log_line:
                logger.info(f"[garak log] {log_line}")
                if self.log_callback:
                    await self.log_callback(log_line)
        
        await process.wait()
        
        if process.returncode != 0:
            error_msg = f"Garak execution failed with exit code {process.returncode}"
            logger.error(error_msg)
            if self.log_callback:
                await self.log_callback(f"🛑 ERROR: {error_msg}")
                await self.log_callback("💡 Tip: If you see a '403 Gated Repo' error (common for Llama models), please use an Open model (like Qwen or Phi-3) or add your GARAK_HF_TOKEN to the .env file.")
            
        jsonl_file = f"{self.output_file}.report.jsonl"
        return await self._parse_garak_output(jsonl_file)

    async def _parse_garak_output(self, file_path: str) -> List[Dict[str, Any]]:
        if not os.path.exists(file_path):
            logger.error(f"[garak_scanner] Report file {file_path} not found.")
            return []
            
        results_map = {}
        
        with open(file_path, "r") as f:
            for line in f:
                if not line.strip():
                    continue
                try:
                    record = json.loads(line)
                    # Garak typically logs 'eval' objects with test results
                    if record.get("entry_type") == "eval":
                        probe = record.get("probe", "unknown")
                        detector = record.get("detector", "unknown")
                        
                        passed = record.get("passed", 0)
                        total = record.get("total", 0)
                        failed = total - passed
                        
                        key = f"{probe}_{detector}"
                        if key not in results_map:
                            results_map[key] = {
                                "probe": probe,
                                "detector": detector,
                                "failed_attempts": 0,
                                "total_attempts": 0
                            }
                        
                        results_map[key]["failed_attempts"] += failed
                        results_map[key]["total_attempts"] += total
                except Exception as e:
                    logger.warning(f"Failed to parse garak JSONL line: {e}")
                    
        # Format into final array
        formatted_results = []
        for v in results_map.values():
            status = "PASS" if v["failed_attempts"] == 0 else "VULNERABLE"
            
            # Simple heuristic for severity if vulnerable
            severity = "info"
            if status == "VULNERABLE":
                pct = v["failed_attempts"] / max(v["total_attempts"], 1)
                if pct > 0.5: severity = "critical"
                elif pct > 0.2: severity = "high"
                else: severity = "medium"
                
            formatted_results.append({
                "probe": v["probe"],
                "detector": v["detector"],
                "status": status,
                "failed_attempts": v["failed_attempts"],
                "total_attempts": v["total_attempts"],
                "severity": severity
            })
            
        return formatted_results

async def fetch_available_models(limit: int = 50) -> List[Dict[str, Any]]:
    """
    Fetches trending, non-gated text generation models from Hugging Face Hub for a 'Direct' experience.
    """
    import httpx
    try:
        async with httpx.AsyncClient() as client:
            res = await client.get(
                "https://huggingface.co/api/models",
                params={
                    "pipeline_tag": "text-generation",
                    "sort": "downloads",
                    "direction": "-1",
                    "limit": limit,
                    "full": "true" # Need this for gated metadata
                },
                timeout=10.0
            )
            res.raise_for_status()
            models = res.json()
            
            # Filter for non-gated models to ensure friction-free experience
            open_models = []
            for m in models:
                if not m.get("gated"):
                    open_models.append({
                        "id": m["id"], 
                        "downloads": m.get("downloads", 0),
                        "gated": False
                    })
            
            return open_models[:35] 
    except Exception as e:
        logger.error(f"Failed to fetch HF models: {e}")
        # Optimized fallback: High-quality Open Models (No token required)
        return [
            {"id": "Qwen/Qwen2.5-0.5B-Instruct", "downloads": 1000000, "gated": False},
            {"id": "Qwen/Qwen2.5-1.5B-Instruct", "downloads": 900000, "gated": False},
            {"id": "microsoft/Phi-3-mini-4k-instruct", "downloads": 800000, "gated": False},
            {"id": "openai-community/gpt2", "downloads": 500000, "gated": False},
            {"id": "HuggingFaceTB/SmolLM-135M", "downloads": 300000, "gated": False},
            {"id": "stabilityai/stablelm-zephyr-3b", "downloads": 200000, "gated": False}
        ]
