hacknova/hacknova_ps-main/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo==4.7.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
celery==5.4.0
redis==5.0.6
python-nmap==0.7.1
python-dotenv==1.0.1
requests==2.32.2
urllib3==2.2.1
networkx==3.3
neo4j==5.20.0
reportlab==4.2.0
sentence-transformers==3.0.1
chromadb==0.5.3
openai==1.30.5
httpx==0.27.0
pydantic==2.7.2
aiofiles==23.2.1
minio==7.2.7
Pillow==10.3.0
matplotlib==3.9.0
